DUI: validate image size before attempting to set as icon [2/2]
Validate all custom icon image sizes before we move on to further
processing. Set a max allowed image size of 512x512. This should
act as somewhat of a "sanity filter" for the users that try to set
wallpapers as DUI icons ;D

ezio: set the limit to 512 (added bitmapfactory sample
scaling in the 1/2 part so we still have small high quality bitmaps
without system performances loss)
fixed fling custom logo img logic so the img is not saved on sdcard
if size>512

Change-Id: I65f31c48e868873eca63695ca15a2495e638d6ee
Signed-off-by: Pafcholini <nadyaivanova14@gmail.com>